Successfully used react router to display and link to the app's different pages
