+++ src/Components/HomeMain.jsx
@@ -9,9 +9,11 @@ const HomeMain = () => {
 
     return (
 
-        <>
+        <main>
 
-        </>
+            <h1>This is the home page.</h1>
+
+        </main>
 
     )
