+++ src/Components/Temp.jsx
@@ -1,0 +1,18 @@
+// UserSearchMain.jsx
+// This is the view that shows all user search results in the current search
+
+const Temp = () => {
+
+
+
+    return (
+
+        <>
+            <h3>This is the user search page</h3>
+        </>
+
+    )
+
+}
+
+export default Temp;
